clk: qcom: apcs-msm8916: call devm_clk_get() with id
With commit fbefb7cc8e51 ("clk: qcom: apcs-msm8916: use clk_parent_data
to specify the parent") in place, it becomes fragile to call
devm_clk_get() with NULL id for getting "pll" clock.  In case that there
are multiple phandles in 'clocks' property, they can be ordered
arbitrarily and identified with 'clock-names'.  That said, "pll" clock
might not always be the first phandle in 'clocks' property.

Let's test the count of clock phandles and use "pll" id for the call, if
there are multiple clocks, and fall on NULL id otherwise.

Signed-off-by: Shawn Guo <shawn.guo@linaro.org>
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>